Footer was separated as independent component
