+++ src/pages/MainPage/MainPage.jsx
@@ -2,7 +2,7 @@ import NavBar from '../../components/common/NavBar/NavBar';
 import HeroSection from './HeroSection';
 import AboutSection from './AboutSection';
 import CitiesSection from './CitiesSection';
-import Footer from './Footer';
+import Footer from '../../components/common/Footer/Footer';
 import BecomeGuide from './BecomeGuide';
 import OurGuidesSection from './OurGuidesSection';
 import ContactsSection from './ContactsSection';
